ПР11
добавлены лайки, дизлайки, удаление карточек

+++ src/utils/Api.js
@@ -60,20 +60,20 @@ class Api {
             .then(this._checkResponse);
     };
 
-    like(id) {
-        return fetch(`${this._url}cards/${id}/likes`, {
-            method: 'PUT',
-            headers: this._headers,
-        })
-            .then(this._checkResponse);
-    };
-
-    dislike(id) {
-        return fetch(`${this._url}cards/${id}/likes`, {
-            method: 'DELETE',
-            headers: this._headers,
-        })
-            .then(this._checkResponse);
+    toggleLikeStatus(id, isLiked) {
+        if (isLiked) {
+            return fetch(`${this._url}cards/${id}/likes`, {
+                method: 'PUT',
+                headers: this._headers,
+            })
+                .then(this._checkResponse);
+        } else {
+            return fetch(`${this._url}cards/${id}/likes`, {
+                method: 'DELETE',
+                headers: this._headers,
+            })
+                .then(this._checkResponse);
+        }
     };
 
     editAvatar(data) {
